Create npm error on ubuntu 18.04
I'm following an online training but the guy is giving instructions for Windows and Mac, so I'm just adding the command lines to install the indicated programs. Git seemed to install with no issues, Visual Studio code gave a warning but seemed to run fine. When I got the npm errors I went away from the instructions and downloaded nodejs which seemed to work. It's just when trying to install npm that just continues to give that error.